refactor: move FilterIncidents from IncidentsList component
